Fix electric field calculation, v1.0 (#3)
* Update `pramp` such that it always solves for the first bias.

* Use one-sided  "upwinded" gradient evolution

Up to now, electric field was evaluated by averaged gradients in the
inner of the domain and by one-sided gradients at the boundries.
This inconsistency led to some mesh size dependent and thus unphysical
concentration jumps in the first and last grid cells.

Now the we always use one-sided evaluation, switching the direction
in the center of the domain. This gives a seemingly physically more
consistent picture.

The issue is semi-solved and should be investigated further.

* Bump version to 1.0

* update README